Version 3.5.9 1. Now the thread doesn't stop. For Hangman/Gift 2. Update JDA/Lavaplayer 3. more fix....
P. S. Prefix don`t work
