+++ src/main/java/events/MessageWhenBotLeaveJoinToGuild.java
@@ -15,20 +15,18 @@ public class MessageWhenBotLeaveJoinToGuild extends ListenerAdapter {
         String idGuildJoin = event.getGuild().getId();
         String nameGuild = event.getGuild().getName();
         String regionGuild = event.getGuild().getRegion().getName();
-
+        try {
         event.getGuild().getDefaultChannel()
                 .sendMessage("Thanks for adding " + "**mego**" + " to " + event.getGuild().getName() + "!"
                         + "\nUse **!help/help/info** for a list of commands."
-                        + "\nCreate text channel with name: `botlog`"
-                ).queue();
-        EmbedBuilder botAddToGuild = new EmbedBuilder();
-        botAddToGuild.setColor(0xa224db);
-        botAddToGuild.setTitle("Guild added a bot");
-        botAddToGuild.setDescription(
-                "Guild id: " + idGuildJoin
-                        + "\nGuild name: " + nameGuild
-                        + "\nGuild region: " + regionGuild);
-        try {
+                        + "\nCreate text channel with name: `botlog`").queue();
+
+            EmbedBuilder botAddToGuild = new EmbedBuilder();
+            botAddToGuild.setColor(0xa224db);
+            botAddToGuild.setTitle("Guild added a bot");
+            botAddToGuild.setDescription("Guild id: " + idGuildJoin
+                    + "\nGuild name: " + nameGuild
+                    + "\nGuild region: " + regionGuild);
             BotStart.jda.getGuildById("779317239722672128").getTextChannelById("779321076424376350").sendMessage(botAddToGuild.build()).queue();
             botAddToGuild.clear();
         } catch (Exception e) {
@@ -39,7 +37,7 @@ public class MessageWhenBotLeaveJoinToGuild extends ListenerAdapter {
             DataBase dataBase = new DataBase();
             dataBase.createTableGuildWhenBotJoinGuild(idGuildJoin);
             dataBase.createTableVoiceWhenBotJoinGuild(idGuildJoin);
-            dataBase.createDefaultUserInVoice("1", "1111111111111", idGuildJoin);
+            dataBase.createDefaultUserInVoice("1", "111111111111111111", idGuildJoin);
         } catch (SQLException exception) {
             exception.printStackTrace();
         }
